+++ CataloguesAnywhere/ModEntry.cs
@@ -1,4 +1,5 @@
 using CataloguesAnywhere.Framework;
+using ChroniclerCherry.Common.Integrations.GenericModConfigMenu;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
@@ -21,8 +22,10 @@ internal class ModEntry : Mod
     /// <inheritdoc />
     public override void Entry(IModHelper helper)
     {
+        I18n.Init(helper.Translation);
         this.Config = helper.ReadConfig<ModConfig>();
 
+        helper.Events.GameLoop.GameLaunched += this.OnGameLaunched;
         helper.Events.Input.ButtonsChanged += this.OnButtonsChanged;
     }
 
@@ -30,6 +33,16 @@ internal class ModEntry : Mod
     /*********
     ** Private methods
     *********/
+    /// <inheritdoc cref="IGameLoopEvents.GameLaunched" />
+    private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
+    {
+        this.AddGenericModConfigMenu(
+            new GenericModConfigMenuIntegrationForCataloguesAnywhere(),
+            get: () => this.Config,
+            set: config => this.Config = config
+        );
+    }
+
     /// <inheritdoc cref="IInputEvents.ButtonsChanged" />
     private void OnButtonsChanged(object? sender, ButtonsChangedEventArgs e)
     {